Retry the database connection
When booting the db container without already having its data volume
at the same time as the application container, the db is not ready for
connections when the application is starting. This would cause the
application to fail. This change adds database connection retrying to
allow the db time to start up.